+++ __docs__/wonder-blocks-form/radio-group.stories.tsx
@@ -4,7 +4,7 @@ import type {Meta, StoryObj} from "@storybook/react";
 
 import {View} from "@khanacademy/wonder-blocks-core";
 import {semanticColor, spacing} from "@khanacademy/wonder-blocks-tokens";
-import {LabelLarge, LabelMedium} from "@khanacademy/wonder-blocks-typography";
+import {BodyText} from "@khanacademy/wonder-blocks-typography";
 
 import {Choice, RadioGroup} from "@khanacademy/wonder-blocks-form";
 import packageConfig from "../../packages/wonder-blocks-form/package.json";
@@ -154,9 +154,9 @@ export const MultipleChoiceStyling: StoryComponentType = () => {
 
     return (
         <>
-            <LabelLarge style={styles.prompt}>
+            <BodyText weight="bold" tag="span" style={styles.prompt}>
                 Select your blood type
-            </LabelLarge>
+            </BodyText>
             <RadioGroup
                 groupName="science-classes"
                 onChange={setSelectedValue}
@@ -230,9 +230,8 @@ FiltersOutFalsyChildren.parameters = {
 
 /**
  * There are specific situations where you might want to use a custom label
- * component instead of using the default `LabelMedium` component. This example
- * demonstrates how to use a custom label component that can be passed in as a
- * prop to the `RadioGroup` component.
+ * component. This example demonstrates how to use a custom label component
+ * that can be passed in as a prop to the `RadioGroup` component.
  */
 export const CustomLabel: StoryComponentType = {
     ...Default,
@@ -251,8 +250,10 @@ export const CustomLabel: StoryComponentType = {
                     justifyContent: "space-between",
                 }}
             >
-                <LabelLarge>Pokemon</LabelLarge>
-                <LabelMedium>(optional)</LabelMedium>
+                <BodyText weight="bold" tag="span">
+                    Pokemon
+                </BodyText>
+                <BodyText tag="span">(optional)</BodyText>
             </View>
         ),
     },