+++ packages/wonder-blocks-tokens/src/theme/semantic/semantic-color-thunderblocks.ts
@@ -542,7 +542,6 @@ export const semanticColor = mergeTheme(defaultSemanticColor, {
             },
         },
     },
-
     status: {
         critical: {
             background: color.red_90,