+++ packages/wonder-blocks-tabs/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @khanacademy/wonder-blocks-tabs
 
+## 0.3.16
+
+### Patch Changes
+
+- Updated dependencies [d8716ab]
+    - @khanacademy/wonder-blocks-tokens@11.2.2
+    - @khanacademy/wonder-blocks-typography@4.2.9
+
 ## 0.3.15
 
 ### Patch Changes